+++ FirebaseAuth/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- [fixed] Fixed a bug where `verifyPhoneNumber` could hit an infinite recursion and crash due to a Swift concurrency compiler bug in some toolchains. (#16416)
+
 # 12.17.0
 - [fixed] Fixed an array out-of-bounds access in `ActionCodeURL` query parsing.
   (#16382)